+++ include/bulk/partitionings/partitioning.hpp
@@ -11,9 +11,6 @@
 
 namespace bulk {
 
-template <int D>
-using index_type = std::array<int, D>;
-
 /** Base class for partitionings over a 1D processor grid. */
 template <int D>
 class partitioning {