+++ test/partitioning.cpp
@@ -17,22 +17,14 @@ void test_partitioning() {
         int s = world.rank();
         int p = world.active_processors();
 
+        BULK_SECTION("Index manipulations") {
+            bulk::index<2> idx{2, 3};
+            BULK_CHECK(idx[1] == 3, "can construct with initializer list");
+        }
+
         auto N = (int)sqrt(p);
         BULK_SKIP_SECTION_IF("Partitionings", N * N != p);
         BULK_SKIP_SECTION_IF("Partitionings", p <= 1);
-
-        BULK_SECTION("Cyclic partitioning to 1D") {
-            auto part =
-                bulk::cyclic_partitioning<2, 1>({5 * p * N, 5 * p * N}, {p});
-            BULK_CHECK(part.owner({p + 2, 3}) == 2,
-                       "compute correctly the cyclic from 2 -> 1 dim");
-            BULK_CHECK(part.local_size(s)[0] == 5 * N,
-                       "compute correctly the extent in cyclic dim");
-            BULK_CHECK(part.local_size(s)[1] == 5 * p * N,
-                       "compute correctly the extent in non-cyclic dim");
-            BULK_CHECK(part.local({p + 2, 3})[0] == 1,
-                       "compute correctly the local index");
-        }
 
         BULK_SECTION("Cyclic partitioning") {
             auto part =
@@ -46,6 +38,18 @@ void test_partitioning() {
             BULK_CHECK(part.global({1, 1}, {s % N, s / N})[0] ==
                            N + (s % N),
                        "compute correctly the cyclic global index");
+        }
+
+        BULK_SECTION("Cyclic partitioning to 1D") {
+            auto part = bulk::cyclic_partitioning<2, 1>({5 * p * N, 5 * p * N}, p);
+            BULK_CHECK(part.owner({p + 2, 3}) == 2,
+                       "compute correctly the cyclic from 2 -> 1 dim");
+            BULK_CHECK(part.local_size(s)[0] == 5 * N,
+                       "compute correctly the extent in cyclic dim");
+            BULK_CHECK(part.local_size(s)[1] == 5 * p * N,
+                       "compute correctly the extent in non-cyclic dim");
+            BULK_CHECK(part.local({p + 2, 3})[0] == 1,
+                       "compute correctly the local index");
         }
 
         BULK_SECTION("Block partitioning") {
@@ -69,8 +73,7 @@ void test_partitioning() {
 
         BULK_SECTION("Block partitioning custom axes") {
             // construct a block partitioning only in the 2nd axis
-            auto part =
-                bulk::block_partitioning<2, 1>({10 * p, 10 * p}, {p}, {1});
+            auto part = bulk::block_partitioning<2, 1>({10 * p, 10 * p}, p, 1);
 
             BULK_CHECK(part.owner({0, 13}) == 1,
                        "compute correctly the block owner");
@@ -103,7 +106,7 @@ void test_partitioning() {
             auto part =
                 bulk::tree_partitioning<2>({10, 10}, 4, std::move(tree));
 
-            BULK_CHECK((part.local_size(0) == std::array<int, 2>{5, 5}),
+            BULK_CHECK((part.local_size(0) == bulk::index<2>{5, 5}),
                        "extent of bspart are correct");
 
             BULK_CHECK((part.owner({1, 1}) == part.owner({2, 2})),
@@ -112,16 +115,15 @@ void test_partitioning() {
             BULK_CHECK((part.owner({1, 1}) != part.owner({6, 7})),
                        "assign correct owner bspart (2)");
 
-            BULK_CHECK((part.origin(1) == std::array<int, 2>{5, 0}),
+            BULK_CHECK((part.origin(1) == bulk::index<2>{5, 0}),
                        "assign correct origin bspart (1)");
-            BULK_CHECK((part.origin(2) == std::array<int, 2>{0, 5}),
+            BULK_CHECK((part.origin(2) == bulk::index<2>{0, 5}),
                        "assign correct origin bspart (2)");
-            BULK_CHECK((part.origin(3) == std::array<int, 2>{5, 5}),
+            BULK_CHECK((part.origin(3) == bulk::index<2>{5, 5}),
                        "assign correct origin bspart (3)");
 
-            BULK_CHECK(
-                (part.local({6, 6}) == std::array<int, 2>{1, 1}),
-                "assign correct origin bspart");
+            BULK_CHECK((part.local({6, 6}) == bulk::index<2>{1, 1}),
+                       "assign correct origin bspart");
         }
 
         BULK_SECTION("Partitioned array") {
@@ -147,14 +149,14 @@ void test_partitioning() {
         }
 
         BULK_SECTION("Irregular block partitioning") {
-          auto part = bulk::block_partitioning<1>({5}, {4});
+          auto part = bulk::block_partitioning<1>(5, 4);
           BULK_CHECK(part.local_count(0) == 2, "large block comes first");
           BULK_CHECK(part.local_count(1) == 1, "small block comes after");
           BULK_CHECK(part.local_count(3) == 1, "last block is small");
-          BULK_CHECK(part.owner({0}) == 0, "computes owner for 0 correctly")
-          BULK_CHECK(part.owner({1}) == 0, "computes owner for 1 correctly")
-          BULK_CHECK(part.owner({3}) == 2, "computes owner for 3 correctly")
-          BULK_CHECK(part.owner({4}) == 3, "computes owner for 4 correctly")
+          BULK_CHECK(part.owner(0) == 0, "computes owner for 0 correctly")
+          BULK_CHECK(part.owner(1) == 0, "computes owner for 1 correctly")
+          BULK_CHECK(part.owner(3) == 2, "computes owner for 3 correctly")
+          BULK_CHECK(part.owner(4) == 3, "computes owner for 4 correctly")
         }